backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pgvector
pydantic>=2.0
python-dotenv
google-genai